update series service validation

+++ series-service/src/main/java/com/edu/service/impl/EpisodeServiceImpl.java
@@ -12,6 +12,7 @@ import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
 import java.util.List;
+
 @Service
 @RequiredArgsConstructor
 public class EpisodeServiceImpl implements EpisodeService {
@@ -21,6 +22,7 @@ public class EpisodeServiceImpl implements EpisodeService {
     private final EpisodeMapper episodeMapper;
 
     private final SeasonRepo seasonRepo;
+
     @Override
     public EpisodeDto getById(int id) {
         return episodeMapper.toDto(episodeRepo.findById(id).get());
@@ -31,29 +33,32 @@ public class EpisodeServiceImpl implements EpisodeService {
         var season = seasonRepo.findById(seasonId).get();
         return episodeMapper.toListDto(season.getEpisodes());
     }
+
     @Transactional
     @Override
-    public EpisodeDto save(int seasonId, RequestEpisodeDto episodeDto) {
+    public EpisodeDto save(int seasonId, RequestEpisodeDto episodeDto, String userId) {
         var season = seasonRepo.findById(seasonId).get();
-        var newEpisode = new Episode(episodeDto.getName(), episodeDto.getEpisodeNumber(), episodeDto.getDuration(), season, "1");
+        var newEpisode = new Episode(episodeDto.getName(), episodeDto.getEpisodeNumber(), episodeDto.getDuration(), season, userId);
         episodeRepo.save(newEpisode);
         return episodeMapper.toDto(newEpisode);
     }
 
     @Transactional
     @Override
-    public EpisodeDto update(int id, RequestEpisodeDto episodeDto) {
+    public EpisodeDto update(int id, RequestEpisodeDto episodeDto, String userId) {
         var updatingEpisode = episodeRepo.findById(id).get();
-        if(updatingEpisode==null) return null;
-        if(episodeDto.getName()!=null) updatingEpisode.setName(episodeDto.getName());
-        if(episodeDto.getEpisodeNumber()!=null) updatingEpisode.setEpisodeNumber(episodeDto.getEpisodeNumber());
-        if(episodeDto.getDuration()!=null) updatingEpisode.setDuration(episodeDto.getDuration());
+        if (updatingEpisode == null || !updatingEpisode.getOwnerId().equals(userId)) return null;
+        if (episodeDto.getName() != null) updatingEpisode.setName(episodeDto.getName());
+        if (episodeDto.getEpisodeNumber() != null) updatingEpisode.setEpisodeNumber(episodeDto.getEpisodeNumber());
+        if (episodeDto.getDuration() != null) updatingEpisode.setDuration(episodeDto.getDuration());
         return episodeMapper.toDto(updatingEpisode);
     }
 
     @Transactional
     @Override
-    public void delete(int id) {
+    public void delete(int id, String userId) {
+        var deletingEpisode = episodeRepo.findById(id).get();
+        if (deletingEpisode == null || !deletingEpisode.getOwnerId().equals(userId)) return;
         episodeRepo.deleteById(id);
     }
 }
